// src/interfaces/IMyPointTranslate.tsx
export interface IMyPointTranslate {
  myRewardPoints: string;
  availablePoints: string;
  processingPoints: string;
  claimVoucher: string;
  seeAll: string;
  pointHistory: string;
  claimNow: string;
  myPoints: string;
  pleaseUseBefore: string;
  youExchangedSuccess: string;
  take: string;
  points: string;
}